Cast sparse node‐feature indices to float
When sparse=True, create_node_features now returns an (N,1) tensor of floats (instead of longs).

+++ test/fixtures/data_fixtures.py
@@ -60,7 +60,9 @@ def create_node_features(
         torch.Tensor: Node features
     """
     if sparse:
-        return torch.randint(0, in_features, (num_nodes, 1))
+        return torch.randint(
+            0, in_features, (num_nodes, 1)
+        ).float()
     return torch.randn(num_nodes, in_features)
